+++ DfE.FindInformationAcademiesTrusts/Pages/Trusts/Academies/PupilNumbers.cshtml.cs
@@ -1,6 +1,7 @@
 using DfE.FindInformationAcademiesTrusts.Data;
 using DfE.FindInformationAcademiesTrusts.Data.Enums;
 using DfE.FindInformationAcademiesTrusts.Services;
+using DfE.FindInformationAcademiesTrusts.Services.Academy;
 using DfE.FindInformationAcademiesTrusts.Services.DataSource;
 using DfE.FindInformationAcademiesTrusts.Services.Trust;
 using Microsoft.AspNetCore.Mvc;
@@ -9,12 +10,15 @@ namespace DfE.FindInformationAcademiesTrusts.Pages.Trusts.Academies;
 
 public class PupilNumbersModel : AcademiesPageModel
 {
-    public Trust Trust { get; set; } = default!;
+    public IAcademyService AcademyService { get; }
+    public AcademyPupilNumbersServiceModel[] Academies { get; set; } = default!;
 
     public PupilNumbersModel(ITrustProvider trustProvider, IDataSourceService dataSourceService,
-        ILogger<PupilNumbersModel> logger, ITrustService trustService, IExportService exportService, IDateTimeProvider dateTimeProvider)
+
+        ILogger<PupilNumbersModel> logger, ITrustService trustService,IAcademyService academyService, IExportService exportService, IDateTimeProvider dateTimeProvider)
         : base(trustProvider, dataSourceService, trustService, exportService, logger, dateTimeProvider)
     {
+        AcademyService = academyService;
         PageTitle = "Academies pupil numbers";
         TabName = "Pupil numbers";
     }
@@ -25,7 +29,7 @@ public class PupilNumbersModel : AcademiesPageModel
 
         if (pageResult.GetType() == typeof(NotFoundResult)) return pageResult;
 
-        Trust = (await TrustProvider.GetTrustByUidAsync(Uid))!;
+        Academies = await AcademyService.GetAcademiesInTrustPupilNumbersAsync(Uid);
 
         DataSources.Add(new DataSourceListEntry(await DataSourceService.GetAsync(Source.Gias),
             new List<string> { "Pupil numbers" }));
@@ -33,7 +37,10 @@ public class PupilNumbersModel : AcademiesPageModel
         return pageResult;
     }
 
-    public string PhaseAndAgeRangeSortValue(Academy academy)
+
+    public string TabName => "Pupil numbers";
+
+    public static string PhaseAndAgeRangeSortValue(AcademyPupilNumbersServiceModel academy)
     {
         return $"{academy.PhaseOfEducation}{academy.AgeRange.Minimum:D2}{academy.AgeRange.Maximum:D2}";
     }